movies_admin/requirements-dev.txt
-r requirements.txt
flake8==6.0.0
django-debug-toolbar==3.4.0